-Wsign-conversion
Fixd a bug:
hv ^= hash_block(&hs, bb_several(b) & prng(&rs));

Intent was to use b' = b & prng(&rs) as input to bb_several().

Unfortunately, bool is implicitly convertible into any integer type, without
complaint, even using -Wsign-conversion. So this wasn't spotted by the compiler.

+++ engine.c
@@ -89,7 +89,7 @@ Engine engine_new(const char *cmd, const char *name, FILE *log, Deadline *deadli
         assert(c);
 
         str_cpy(&line, "setoption name ");
-        str_ncat(&line, token.buf, c - token.buf);
+        str_ncat(&line, token.buf, (size_t)(c - token.buf));
         str_cat(&line, " value ", c + 1);
 
         engine_writeln(&e, line.buf);
